部分接口对接

+++ src/components/Home.vue
@@ -2,24 +2,31 @@
   <div class="home-main-container">
     <div class="scrollsnap section1">
       <img class="section1-logo" src="../assets/title_ver2.svg" alt="" />
-      <div class="section1-msg">立即开始</div>
+      <div class="section1-msg" @click="flipActive = true">立即开始</div>
       <div class="section1-login-frame">
-        <div class="login-form">
-          <p>
-            用户名
-            <input v-model="formName" type="text" />
-          </p>
-          <p>
-            密码
-            <input v-model="formPassword" type="text" />
-          </p>
-          <p>{{ formMsg }}</p>
-          <Button @click="login">登录</Button>
+        <div :class="{ 'flip-frame': true, 'flip-active': flipActive }">
+          <div class="front"></div>
+          <div class="back login-form">
+            <p>
+              用户名
+              <input v-model="formName" type="text" />
+            </p>
+            <p>
+              密码
+              <input v-model="formPassword" type="text" />
+            </p>
+            <p>{{ formMsg }}</p>
+            <Button @click="login">登录</Button>
+            <Button @click="register">注册</Button>
+          </div>
         </div>
       </div>
     </div>
-    <div class="scrollsnap section2"></div>
+    <div class="scrollsnap section2">
+      <div></div>
+    </div>
     <div class="scrollsnap section3"></div>
+    <div class="scrollsnap section4"></div>
   </div>
 </template>
 
@@ -28,28 +35,37 @@ import Button from "./Button.vue";
 import { useStore } from "../store";
 import { useRouter } from "vue-router";
 import { ApiLogin } from "../api";
-import axios from "axios";
 import { ref } from "vue";
 
 const router = useRouter();
 const store = useStore();
-
-const formName = ref("");
 const formPassword = ref("");
+const formName = ref("");
 const formMsg = ref("");
+
+const flipActive = ref(false);
 
 const login = () => {
   console.log("login function:");
   if (formName.value && formPassword.value) {
     ApiLogin(formName.value, formPassword.value)
       .then((res) => {
-        const { userToken } = res.data;
-        console.log(res.data);
-        localStorage.setItem("token", userToken);
-        store.isLogin = true;
-        router.push({
-          path: "/user",
-        });
+        console.log(res);
+        const { userNameExist, userPasswordOk, userToken } = res.data;
+        if (!userNameExist) {
+          formMsg.value = "用户名不存在";
+        } else if (!userPasswordOk) {
+          formMsg.value = "密码错误";
+        } else {
+          localStorage.setItem("token", userToken);
+          localStorage.setItem("userName", formName.value);
+          store.isLogin = true;
+          store.token = userToken;
+          store.userName = formName.value;
+          router.push({
+            path: "/user",
+          });
+        }
       })
       .catch((res) => {
         console.log(res);
@@ -58,16 +74,19 @@ const login = () => {
     formMsg.value = "※请输入用户名和密码";
   }
 };
+
+const register = () => {};
 </script>
 
 <style lang="less" scoped>
 .home-main-container {
   width: 100%;
   height: 100%;
+  background-color: #fff;
   overflow-y: scroll;
 
   .scrollsnap {
-    max-width: 1366px;
+    // max-width: 1366px;
     height: min(56vw, 758px);
     margin: 0 auto;
   }
@@ -97,10 +116,43 @@ const login = () => {
       align-items: center;
       justify-content: center;
 
+      .flip-frame {
+        height: 100%;
+        width: 100%;
+        max-width: 500px;
+        max-height: 600px;
+        border-radius: 10px;
+        position: relative;
+        transform-style: preserve-3d;
+        transition: 300ms;
+
+        .front {
+          position: absolute;
+          width: 100%;
+          height: 100%;
+          backface-visibility: hidden;
+        }
+
+        .back {
+          background-color: #eee;
+          width: 100%;
+          position: absolute;
+          backface-visibility: hidden;
+          height: 100%;
+          transform: rotateY(180deg);
+        }
+      }
+
+      .flip-active {
+        transform: rotateY(180deg);
+      }
+
       .login-form {
         background-color: #eee;
         height: 100%;
         width: 100%;
+        max-width: 500px;
+        max-height: 600px;
         border-radius: 10px;
         padding: 10px;
         display: flex;
@@ -109,6 +161,10 @@ const login = () => {
         justify-content: space-around;
       }
     }
+  }
+
+  .section2 {
+    background-color: #eee;
   }
 }
 </style>
